Moved code to ThemedActivity from PreferenceActivity
