+++ app/src/main/java/com/amaze/filemanager/ui/activities/superclasses/PreferenceActivity.java
@@ -65,27 +65,6 @@ public class PreferenceActivity extends BasicActivity {
     sharedPrefs = PreferenceManager.getDefaultSharedPreferences(this);
   }
 
-  @Override
-  public void onConfigurationChanged(@NonNull Configuration newConfig) {
-    super.onConfigurationChanged(newConfig);
-
-    if (getPrefs().getString(PreferencesConstants.FRAGMENT_THEME, "4").equals("4")) {
-      getUtilsProvider().getThemeManager().setAppTheme(AppTheme.getTheme(this, 4));
-      restartPC(this);
-    }
-    }
-
-  public static void restartPC(final Activity activity) {
-    if (activity == null) return;
-
-    final int enter_anim = android.R.anim.fade_in;
-    final int exit_anim = android.R.anim.fade_out;
-    activity.overridePendingTransition(enter_anim, exit_anim);
-    activity.finish();
-    activity.overridePendingTransition(enter_anim, exit_anim);
-    activity.startActivity(activity.getIntent());
-  }
-
   public SharedPreferences getPrefs() {
     return sharedPrefs;
   }
